MCOL-1101 Move client UDFs into a separate file. 	Remove rmParms from ha_calpont_impl.cpp
